+++ CHANGELOG.rst
@@ -4,7 +4,11 @@ Changelog
 
 Changes:
 
-v0.0.1
+v0.0.2 (futuru)
+  - add heuristic methods
+  - add dataset type with subsample generator
+
+v0.0.1 (stable)
   - added simple example
   - added documentation
   - added statistical methods for sample size estimation